fix: use wrapper to load sw
Uses a small wrapper file to load the main service worker. The wrapper
contains the git commit the project was built from so browsers checking
for updates only need to download this much smaller file to know
something has changed and a new version needs to be installed.

As a side-benefit users self-hosting on a domain not on the public
suffix list  will load a (potentially) cached version of the service
worker instead of downloading the full bundle from each subdomain every
time.

Refs: #1146
